+++ concert_simple_scheduler/CHANGELOG.rst
@@ -4,4 +4,5 @@ Change history
 0.7.0 (Forthcoming)
 -------------------
 
- * Simple example concert scheduler node written in Python.
+ * Release to Indigo.
+ * Example concert scheduler node written in Python.